requirements.txt
google-api-python-client==1.5.5
google-cloud==0.21.0
google-cloud-datastore==0.21.0
google-cloud-speech==0.21.0
google-cloud-storage==0.21.0
google-cloud-translate==0.21.0
oauth2client==3.0.0
python-telegram-bot==5.2.0
google-api-python-client==1.5.5
